+++ lib/provider.dart
@@ -1,7 +1,10 @@
 library provider;
 
 export 'src/async_provider.dart';
+export 'src/change_notifier_provider.dart';
 export 'src/consumer.dart';
 export 'src/delegate_widget.dart';
 export 'src/listenable_provider.dart';
 export 'src/provider.dart';
+export 'src/proxy_provider.dart' hide NumericProxyProvider, Void;
+export 'src/value_listenable_provider.dart';