+++ lib/src/async_provider.dart
@@ -1,8 +1,9 @@
 import 'dart:async';
 
 import 'package:flutter/widgets.dart';
-import 'package:provider/src/delegate_widget.dart';
-import 'package:provider/src/provider.dart';
+
+import 'delegate_widget.dart';
+import 'provider.dart';
 
 typedef ErrorBuilder<T> = T Function(BuildContext context, Object error);